Add feature to check a diskette for pre-formatted (factory formatted) sectors.
The heuristic assumes that if the all the sectors for a given side
of track 0 all read good and they all have the same CRC value, that
that side of the disk has no useful data.

If both sides are found pre-formatted, the read stops.  If side 1 is
found pre-formatted, it restarts reading only side 0.

The behavior of the check can be controlled with the -b option.