Change to doing an index to index read when hole detection (-h 1) is enabled.  (It is enabled by default.)
This change is necessary to prevent data loss due to a race window
between detecting the index hole and enabling an unconditional read.
This also solves the infamous problem reported by Ira where, if the
timing is just wrong, cw2dmk can just lose an entire sector of a
track.

The unconditional read is still used with "-h 0".  It should work
the same as before since there is no race window in that code path.