Compile the crc test app
I still think of the crc app as just a developer tool, so I didn't add
it to "progs"; you have to do "make crc" to compile it.  I did add
some minimal documentation in a comment.

I also updated crc.c and the Makefile so that I can compile the
package without warnings on both Ubuntu 20.04 (gcc version 9.3.0) and
Ubuntu 8.04 (gcc version 4.2.4) without warnings.  The crc.c change
was needed for gcc 9.3.0, and the Makefile change (-std=gnu99) was
needed for 4.2.4, which otherwise enforces an even older C standard.