More minor cleanup
It doesn't make sense for the usage message to show a default value
for -j, because -j doesn't take an argument, so you can only use it to
turn the feature on, not off.  Thus "off" obviously is the default.

Also, all other default values are shown in square brackets, but this
one was in parentheses.