Quick fix for DMK "number of tracks" field after replay
Replay currently doesn't update the number of tracks field in the DMK
header to the number of tracks replayed, so initialize it to a value
that fits in a byte (MAX_TRACKS, which is 88) instead of 1000000.

It would be better to merge (or at least use similar logic) with the
guess_tracks feature.  Not only does that feature set the number of
tracks field, but it has logic to avoid writing a garbage track after
the last good track.

For now, to get a clean replay with only the correct number of tracks,
include the -t option instead of letting it default.