Better fix for DMK "number of tracks", plus related cleanup
Make the logic and messages for guessing the number of tracks and
sides more similar between original read and replay.

In particular, the change to number of tracks logic results in the DMK
file containing the correct value.

Guessing sides still isn't quite the same between original read and
replay.  On original reads, it can make sense to try reading the back
even if track 0 appears unformatted or has no good sectors, because in
theory other tracks could exist.  On replay, it doesn't make sense to
try reading the back if the log being replayed has no data for the
back.

Got rid of restarting an original capture when the back has no good
sectors.  It works to simply set sides = 1 and update the dmk header
flags.  Replay already was doing it that way.

Added newlines to the end of some messages that always precede
starting a new track.

Testing:

Compared the .dmk output between the original read and a replay.
(This actually revealed another unrelated bug that I'll fix in the
next commit.)

Did an original read of a disk that is detected as formatted but
with no good sectors on the back, to make sure that the change to that
logic works.