Fix some bugs/omissions in log parsing
Parse '{' and '}' in the sample stream.  If index hole sensor capture
was on, '{' means that subsequent samples had the high-order bit set,
while '}' means subsequent samples no longer have the bit set.

Even if index hole sensor capture is off, due to a bug in cw2dmk.c's
logging, the sample end marker 0x80 is erroneously logged as '{'
followed by a 0-valued sample.  This change effectively works around
that bug too.  (It would be good to fix the logging bug in the future,
but it's good to have the workaround anyway so that we can replay logs
produced by older versions of cw2dmk that still have the bug.)

Fix a bug where something like "2 good sectors, 0 errors" after the
end of the samples on a track could be interpreted as samples with
values 2 and 0.  The fix is to tighten the scanf pattern so that we
can detect when a number isn't immediately followed by s, m, or l.

Make the test program built into parselog slightly more user-friendly,
by giving an error message instead of a core dump if there is no such
file.

Testing:

Replayed decoding a disk.  Compared the original DMK and the replayed
DMK; they are identical now.  Also compared the original logfile and a
-v7 logfile from the replay.  They are similar except for expected
message differences.  In particular the samples parsed out of the
replay log are exactly the same as parsed out of the original log.