Disable retry limit in replay mode
Prior to this change, -x would have to be given during replay mode in
order to use more than the default 4 retries per track from the log
being replayed, even if the log had more retries.  This behavior
doesn't seem at all useful, so this change allows all retries in the
log to always be used.  Replay mode will still stop using more retries
if the track is read without error and -X was not given.

This change also disallows giving the -x option in replay mode, since
it would be ignored, similarly to how other options that aren't
applicable in retry mode are handled.

Testing: Tested with a log from Ira that has a large number of retries
of a bad track.