Fix bug in replay single sided detection
The log was captured single sided if the *first* time we look for
track 0, side 1, we get track 1, side 0.  Add missing test that this
was the first time; otherwise we erroneously detect single sided when
looking for a retry of track 0, side 1.

Testing: Tested with an old double-sided capture.  I'm not sure why I
missed this bug earlier.