Allow -h0 in replay mode and document the need for it
Reading in -h0 has the extra feature (previously not documented in the
man page) of starting the DMK track 48 bytes prior to the first IDAM
decoded from the track.  This feature exists because otherwise there
would often be so much garbage (i.e., a partial sector) at the front
of the DMK track that the DMK track buffer would fill up before there
is room for all the valid sectors.

This feature is needed when replaying an -h0 capture too, so this
change allows that and recommends it in the man page.

Future: It might be a good idea to read the original options that were
used in making the capture from the log file, and use at least some of
them as defaults for the replay.  -k and -h are prime candidates.