Cleanups after review comments
secsize() is now moved to secsize.c and shared by cw2dmk and dmk2cw,
but each one uses separate definitions for encodings (FM, MFM, etc.),
so it is #included instead of separately compiled.  In the future,
perhaps this can be cleaned up further and more code can be shared.