Improve ChangeLog comment about QUIRK_MFM_CLOCK
This quirk was needed to read some older UDOS disks with clock bits
set in certain places that violate the MFM clocking rules.  The most
frequent violation seen on a sample disk was that in a sequence of
bytes with data value ff ff ff..., the clocking pattern could be 00 88
88... instead of 00 00 00...  This results in flux transitions in
adjacent bit cells, which is normally illegal in MFM.  The disks were
probably from UDOS 1526 V.3, or possibly V.2.