Fix two bugs in sector stats
Fixes #23

1) Update enc_sec[] for the current sector from FM to RX02 when the
DAM is seen and the sector is discovered to be RX02.  This bug had
only a rather subtle effect: it caused RX02 sectors that were reused
from an earlier track read by the -j feature to be counted as FM in
the stats (issue #23).

2) When reusing a sector from a previous read, update status using the
encoding that was discovered for it on that read, not the one
discovered for it on the most recent read (which had an error).  This
bug would have very rarely if ever been noticeable, since even on
erroneous sector reads, the encoding is nearly always discovered
correctly.