Add partial support for RX02 in dmk2jv3
Instead of erroring out when given an RX02 .dmk file, produce a .jv3
file that records the double density RX02 format sectors as
conventional MFM sectors.

I found this change useful to check the validity of an RX02 .dmk file
produced by cw2dmk, as dmk2jv3 analyzes the dmk file, finds all the
sectors, and checks all the CRCs.  I didn't really make any use of the
.jv3 output, but I did check that jv2dmk and a couple of other tools
can parse it as intended -- that is, they see a valid .jv3 file with
the double density sectors recorded as conventional MFM of the correct
size (twice what the original sizecode implies for FM).